Jana: Reorder Show Command Popups toggle, unify new VoiceCommandManager logic for disabling alerts when popups are off.
